+++ src/utils/accountUrl.ts
@@ -1,0 +1,3 @@
+export function buildAccountDetailsUrl(accountId: string): string {
+    return `https://digital.americafirst.com/americafirstdigitalbanking/uux.aspx#/account/${accountId}?currentTab=transactions&returnTo=Home`;
+}