+++ src/utils/logger.ts
@@ -1,0 +1,14 @@
+import { debugLoggingSetting } from './settings';
+
+let debugLoggingEnabled = false;
+
+debugLoggingSetting.getValue().then((value) => {
+    debugLoggingEnabled = value;
+});
+debugLoggingSetting.watch((value) => {
+    debugLoggingEnabled = value;
+});
+
+export function debugLog(...args: unknown[]): void {
+    if (debugLoggingEnabled) console.debug(...args);
+}